% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{format_excel_otutable}
\alias{format_excel_otutable}
\title{A function to format a phyloseq object as an otu table for excel}
\usage{
format_excel_otutable(physeq, fname = FALSE)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{fname}{An optional filename/path. Saves to file if TRUE.}
}
\description{
A function to format a phyloseq object as an otu table for excel
}
\examples{
format_excel_otutable(myphyseq, fname = "otu_table.csv")
}

